+++ sciety_labs/app/app_providers_and_models.py
@@ -21,7 +21,7 @@ from sciety_labs.providers.article_recommendation import (
 from sciety_labs.providers.async_providers.crossref.providers import (
     AsyncCrossrefMetaDataProvider
 )
-from sciety_labs.providers.async_providers.opensearch.async_opensearch_article_recommendation import (  # noqa pylint: disable=line-too-long
+from sciety_labs.providers.async_providers.opensearch.providers import (  # noqa pylint: disable=line-too-long
     AsyncOpenSearchArticleRecommendation
 )
 from sciety_labs.providers.async_providers.semantic_scholar.providers import (
